1 February 2010 This is a summary of the changes I made in build 93. It basically consisted of a rewrite of the puddetag GUI.
1. Adding any more to puddleag.py just increased complexity, because a) there was a bunch of connections that had to be tracked and b)functions were called willy-nilly. This resulted in having me having to trace the code anytime I wanted to add/change anything. Obviously, I wasn't pleased with that.

2. Mappings were fucked up. Everything that displayed/wrote tags had to be aware of mappings to use them. Not cool.

3. A lot of the code wasn't really needed anymore. There were better ways of doing shit.

So, while it worked. I'm sure I would've broken something irrevocably in the future. A couple of changes were made.

puddlestuff/puddletag.py was left as is and a new maintest.py module was created. This module is responsible for the mainwindow and its related children.

Controls are now created as follows:
  A module defines the class inheriting from QDialog.
  Three properties are needed: emits, receives and settingsdialog.
  Emits is the names signals that the controls, receveives the signals to be received with corresponding slot.
  These mofos are then connected to appropriate controls in the maintest.
  Hence the addition of tagpanel.py

Another new thing is the PuddleStatus object which is passed to controls upon creation. With this they can query various things like the selected files, pattern text, etc.

And actions/menus are new created dynamically using the loadshortcuts module. I have to say I like it.

In addition the audioinfo libraries were updated to take advantage of tag mappings. I still don't like using a global variable to do it, but I gotta do what I gotta do.