Added an option to change the display format of albums in tag sources. Also made them sortable by fields.
Added a previewMode to the TagModel class in tagmodel.py. This is to allow one to edit retrieved tags using puddletag tag without writing them. Had to modify a lot of methods (especially in puddletag.py) that called preview code in TagModel to take note of this.

Fixed the rename_files function in files.py. It was reading files instead of using the already retrieved tag. Which meant that if an error occurred, a message box wasn't shown.

Added a config menu for allmusic sources, and changed the code so that the configure button is hidden when the tag source doesn't have configuration options.