Changed the get method in audioinfo.util so that it works now as the normal python dictionary get. Also modified the keys method so that it's slightly better to read.
The changes to puddleobjects, and tagmodel are because of the issue where after loading, two files would get highlighted. This was because the setCorner method was called when the model was reset. Also, changed the header so that it sorts automatically instead of calling the sortByColumn method. This is slightly faster.