strip() lines when reading the `genres` config file
Without that, load_genres() of this mock-up file:

```
line1
line2
line3
```

and a subsequent save_genres() would produce:

```
line1

line2

line3

```

At the second execution of puddletag, load_genres() + save_genres() will
generate:

```
line1

line2

line3

```

and so on and so forth. Give it enough puddletag execution and the file would
group in the hundreds of MB, making it impossible to start puddletag again since
it the memory needed to read such a file (made up mostly of empty line) will be
above what's available on the machine.