Fix tagmodel signal connections
The bug was introduced with commit 26eda69b. Calling disconnect() on an
QObject disconnects all slots on all signals, and in this case prevents
any updates the the button states when files are loaded or selected.
It is unfortunate that we can't check if that slot is connected, so we
just have to try and catch the exception in case it is not.

fixes #2