Fix multi-signature signals
Every signature of a signal is effectively a complete signal on its own
with its own slots and emits only to these slots. This does not pair well
at all with the emits/receives auto-connect architecture in puddletag.
Luckily, in all cases we can turn them into single-signature and let
python handle the type magic.

fixes #9