Fix column title display
When returning a UnicodeMod instance as column header title, (py)qt
does not display it properly for some reason. So we always turn it
into a proper str.
Also, changing the language does not change the column titles, because
the titles are saved as-is in the config and can be arbitrarily changed
by the user in the settings dialog.