Drop six module usage
Continuing with removing python2 support, let's drop the usage of the six
module.

most of it was done by hand, following the upstream documentation at:

  https://six.readthedocs.io/

regarding what's the Python3 value for the six compatibility layer.

Given the high number of changes, there's a chance some of them introduced a
regression.