Update the ID3 compat layer
This compat layer was copied as-is from Picard [0] and added write
support for id3v2.3 tags back in the days when mutagen itself did not.
They removed it in 2016 [1] when mutagen itself added support for it.

The here used version caused a filesave to always prepend a completely
new ID3 tag to the file, resulting in files having multiple (stacked)
ID3 tags. "Remove tag" in puddletag caused the first (or last-saved) tag
to be removed, reloading the file then showed the then-first tag (the
previous state). This can be easily verified when looking at a hexdump
of the file. Using "Remove tag" multiple times removes all of these
tags in order until none is left.

I simply replaced it with a current version [2] of the compat layer,
which only adds support for some id3v2.4 frames in id3v2.3. It may be
worthwhile to drop it completely.

[0] https://github.com/metabrainz/picard/blob/a029f36899e19591a9637ec67fb1f7772076014f/picard/formats/mutagenext/compatid3.py
[1] https://github.com/metabrainz/picard/commit/6ed354406adf72ef7972b7f52a4aa94d68a295ad
[2] https://github.com/metabrainz/picard/blob/master/picard/formats/mutagenext/compatid3.py

fixes #51
maybe also #20