Action: to navigate to parent folder of currently selected file
Added button (and menu item), which cause the parent folder of the currently selected file to be selected in the filesystem view. This is useful particularly when opening standalone .mp3 files, as sometimes one wishes to copy tags from other files in the same folder, and navigating to it is a nuisance.