Fix for read-only file crash
Added import for MutagenError to catch the exception.
With these changes no tags are written to the file in the
event of a mutagen error, but the trace is still output to
the console.

A nice addition would be a dialog warning that the files are
read-only when trying to write tags. (I am not versed in PyQt,
otherwise I would do it myself.)

Fixes issue: #546