Allow for multiple files and directories
If you wished to read several files with Puddletag, you previously had
to start the GUI and load the files. If you wrote more files or
directories in the terminal as arguments for Puddletag, only the first
argument would be loaded.

This change will specify from the terminal which files and directory,
and all files specified will be read.