Fix deprecated regex escape sequences in strings
In support of the ongoing adoption of Unicode within Python, [bug # 27364
(2016)][1] deprecates "invalid" (read: regular expression) escape
sequences outside the use of raw strings in Python 3.6+.

This was discovered when performing Flake8 analysis on the code base and
discovering a number of [`W605`][2] warnings thrown.

Due to the fact that, eventually, this will transition from
`DeprecationWarning` to `SyntaxError` this change set resolves all
outstanding `W605` warnings with no other changes made.

Subsequently, as a number if issues have been raised which involve
encoding issues with Unicode and marshalling between strings and bytes
(e.g. #541, #551, #556, #595) some of these regular expression strings
may _also_ need to be changed to [byte literals][3] (i.e. prefixing the
string with `b`).

[1]: https://bugs.python.org/issue27364
[2]: https://www.flake8rules.com/rules/W605.html
[3]: https://docs.python.org/3/library/stdtypes.html#bytes