m3u encoding
- Read m3u files in latin9 encoding, any other with utf-8
- Change the default file ending for exporting from m3u to m3u8

m3u commonly uses latin1/latin9 encoding, which results in
csv.reader() to fail when unicode characters (ä/ö/ü/...) are contained.
The unicode equivalent file ending is called m3u8.

While the default for new exports will be changed for new installations,
unfortunately the legacy name (puddletag.m3u) will not be overridden.
This might cause issues when you will export a playlist as m3u despite
the utf8 suffix and try to reimport it.