Fix tagsources get_encoding
This method leans on the html.parser module and that by specification only accepts strings:

https://docs.python.org/3/library/html.parser.html#html.parser.HTMLParser.feed

It cannot therefore be sensibly used to divine the encoding of a byte-string and this whole block of code is  redundant. Likely a carry over from the Python 2 to 3 conversion.

Moreover I can find no reference in nay documentation to a FoundEncoding exception and no idea where that is implemented but certainly not by html.parser.