Fix bug in AllMusic tagsource
get_encoding returns a string describing the encoding. To overwrite page with that is a glaring bug. Moreover the BeautifulSoup parser doesn't even care about the encoding so we don't need to divine (or if it does it works it out for itself)..

Removing this one line makes the AllMusic tagsource functional once more.

Fixes: https://github.com/puddletag/puddletag/issues/666