An attempt to document the tagsource interface
And provide a class from which tag sources should derive.
Assembling this raises some interesting questions about puddletags interface to tagsources and I suspect it can be improved significantly

Two areas of grave concern:

1. The `search()` method has a very bizarre interface that depends upon the definition of group_by. This should be standardised. Easy to do by fixing `webdb.py`.

2. The group_by attribute seems flexible but there are rather heavily hardcoded assumptions that it's first entry is "album" to my mind,  and the retrieve() method is always receiving the value of the tag that is the first entry in group_by. 

And areas of mild concern arise:

1. `TAGLIST` is unimplemented and no clue is easily found what was envisioned here.
2. `search` and `keyword_search` should really be one function I see no reason to separated these. A simple signature of a `search(`)` method is not hard to devise that can serve both needs, for example:

`def search(text='', files=None, album=None):`

where the arguments are all optional and can be used in a named fashion. And so search is always provided with contents of the search box and list of selected files and optionally an album release name and the tags source implementer can do what they wish with that.