Simplify/rationalise rate limiting of requests
This one puzzles me deeply. How and when did who justify creating a class with an attribute to store a single time value to drive rate limiting? The idea of rate limiting requests to no more than one per second is sound. But while great effort was given to an unneeded class, a configurable rate was not. That is now available.
Fixed a small bug