Clear resource.py
This is just to get rid of the pyrcc5-generated, megabyte-sized binary
content. It will be soon filled with code that more or less acts as a
drop-in replacement.
The `resource.qrc` is intentionally kept for now, as a kind of TODO
list. It will be removed once everything is migrated.