Migrate config data
This is basically implementing the solution suggested here [0]. By
putting the search path additions in the existing `resource.py`, it acts
as a kind of drop-in replacement: if you import that, you will be able
to access the resources using their prefix and path.
The config files here use the superfluous `./` in the path. The reason
for that is that puddletag does some path-magic/-foo using `os.path` which
relies on the fact that the path of the resource have a `/` right before
the actual filename. I don't want to touch that right now, so using this
workaround to keep it working for now.

[0] https://stackoverflow.com/a/66104687/1094486