Update translations
This is just the result of running `update_translations.py`. It is a
separate commit because the translation are changed *a lot*, and in
some cases existing translations are thrown out because the source
changed slightly. If need be reverting this commit will restore the
old translations.
There are now 704 source texts (aka things that can be translated).
Of them ~650 already existed, and ~50 are new (varies between
languages). 12 source texts were obsolete and were removed from the
`.ts` files.