Generate `.qm` files from `.ts` files on build
The `.qm` file are a binary form of the `.ts` file, that is used during
runtime. Using the new build command (which is also executed during a
full build) this is done with QT's `lrelease` command.
Since the translations only works if the `.qm` files exist, this change
has the (somewhat not so nice but bearable) side effect that you have to
run the `build_qm` setuptools command (`build` also works) at least once.
In theory we could also pass the `puddletag.pro` file to the
`lrelease-pro` command to achieve the same result. But the `.pro` file
may not exist, so passing the `.ts` files directly is the safest bet.

This partially reverts 5b70625596db296661390d63888ad5e2b790812e.

See [0] for more info on internationalisation with PyQt5.

[0] https://www.riverbankcomputing.com/static/Docs/PyQt5/i18n.html