Update `puddletag.pro`
Instead of doing partial updates, just rewrite the whole `.pro` file
from the file tree. And since it is only needed for translation updates
and can be generated, there is no need to keep it in git.
Strictly speaking, we don't need the `.pro` at all; we could just pass
the `.py` and `.ts` files directly to `pylupdate5`. In fact, PyQt6
does not support passing `.pro` files to `pylupdate6`. But in PyQt5 the
`.pro` route works better, and the code is a good stepping stone when
the upgrade to PyQt6 comes.