Update puddlelogo
- Moved the svg from the root to the rest of the images
- Resized it from weird 464x127 to not-as-much-weird 512x128 pixel
- Named the layers and objects more descriptively
- Removed the transparency from the background gradient while keeping
  the resulting colours the same. Also resized it to fit the new size,
  and made the white border consistently 2px all around
- The font used for the app name is `Impossibilium BRK` by Ænigma Fonts
  (Brian Kent). It can be found freely on the internet, for example
  here [0]
- Created a path copy so you don't need the font installed
- Re-centered the text
- Updated, repositioned and hid the version text (since it will be
  outdated soon anyway)
- Re-exported the svg to png

In theory we could use the svg directly, but the svg renderer in Qt does
not support filters [1], which are used to create the backdrop shadow.
And since it looks ugly/weird without it, we continue to use a png
version.
If we ever going to use the svg, we could potentially inject the current
version number into it to automatically keep the splash screen
up-to-date.
Did not run `pyrcc5` to regenerate `resource.py` since that will be
replaced soon anyway.

[0] https://www.1001fonts.com/impossibilium-brk-font.html
[1] https://bugreports.qt.io/browse/QTBUG-9081