Fix UFID in ID3
Through some python 2 migration confusion (str/unicode vs. bytes/str)
the handling of the ID3 UFID frames broke. As per the spec [0], these
frames have a binary content (so not necessarily a string). Mutagen
passes the content correctly as `bytes`, but puddletag internally
handles it as a utf8-encoded string, so the id3 wrapper has to handle
the conversions in both directions. If the content should be non-utf8,
it will simply be ignored. This behaviour is not ideal, but I couldn't
think of a better one right now.
With the conversion working again, the stored tags windows only gets
strings again, and puddletag does not crash.

This partially reverts commit c0504240a2b947da91c1e6549ba48040238d8460.

Fixes #590
Fixes #565
Fixes #595

I haven't pinned down the exact versions, but older versions of the
dependencies (like pyqt) behave differently. Instead of crashing the
whole process, they only fail a thread and the process continues to run.
This leaves it in a somewhat undefined state and results in what is
described in #615, where all tags beyond the offending UFID are cut off
in the stored tags. With the UFID fixed, all tags are now displayed.

Fixes #615

[0] https://id3.org/id3v2.4.0-frames   chapter 4.1