Fix hanging thread
Even though this issue does not occur anymore with the UFID fix, it is
still worth fixing. The issue seems to be that the `_loading` flag is
not reset when an error occurs in the `_load(tags)` method, and through
that locking the another thread in the `QApplication.processEvents()`
loop. To always reset that flag, wrap the set und unset of it in a
try/finally block.

Fixes #568
Fixes #647