MusicBrainz: ignore track-list
Sometimes (but not always) a release (which can be a album or a single
or sumsuch) from musicbrainz contains a track-list, which is a list
with some basic info about all the tracks. When writing from the tag
source to files, this release info gets merged with track infos to form
the actual tags to be written. On this way the track-list also gets
passed on and mapped into the files as `track-list` tags with bogus
values.

```
  ...
  'title': 'Stuck on the Puzzle',
  'track': '5',
  'track-list': { 'count': '6',
                  'offset': '0',
                  'track':  [ { 'id': '9d9d60dc-2e9f-3852-954e-7651bc6594d3',
                                'length': '53000',
                                'number': '1',
                                ...
```

The track info from the release is a subset of the actual track info,
and it is never used; so we can just ignore/remove it and be done.

Fixes #659