Fix update_translation.py
It had broken code from the py2 to py3 conversion (attempting an illegal seek on a r+ opened file which bombed, because of the move to unicode as a default in py3).

Project files are small, so it's totally fine reading it in and rewriting it which is what it does now.

Reference to a QT4 utility when puddletag is now QT5.