Fix backslash replace
`re.sub()` is used for the actual replacement, and if its `repl`
argument "is a string, any backslash escapes in it are processed." [0]
This is the reason why any single backslash in the replacement we get
from the gui has to be replaced with two, because two baslashaes will
turn back into one after that processing.
It is a false believe that raw string literals don't do/support *any*
backlash escapes, because they do [1] (last paragraph). Furthermore,
"a raw literal cannot end in a single backslash" [1], so the only way
to have a string literal consisting of only a single backslash is by
using a non-raw one with the valid `\\` escape sequence.
Since `r'\\'` is equal to `'\\\\'`, once could write that as
`replace('\\', r'\\')` but that looks just confusing to the uninitiated.

[0] https://docs.python.org/3/library/re.html#re.sub
[1] https://docs.python.org/3/reference/lexical_analysis.html#string-and-bytes-literals