Use valid fallback lang for id3 uslt
The language of a USLT frame has to be a valid ISO 639-2 language code.
Instead of falling back to `XXX` which is invalid, use the proper `und`
(undefined).
Because of the way puddletag handles multi-value frames as a single
|-sepeared string, it is hard to do a proper validation in the gui.
The language is required for USLT and mutagen will complain harshly
with a crash when it is missing or too short (less than 3 chars). So
extend the fallback to also cover that case.