Preserve column sizes
When column layout changes, take a snapshot of the current column
sizes and apply them to the new layout. Columns are matched by their
name/title, so two columns with the same name/title will both get the
same size of the previously rightmost one of them. Hidden columns have
a size of 0, so be sure to ignore them, elsewise un-hidding won't work.
For obvious reasons, this only applies if puddletag is not configured
to autoresize the columns.
This change has a (IMHO nice) side effect: When you switch from
autoresize to not-autoresize, the columns will retain their last
(autoresized) size and can be manually resized from there.

Fixes #640