Fix changeset detection
With 7f1fa4f2 the `puddlestuff` package was moved and as a result its
relative path/position to the git root changed. The changeset detection
has been updated to accommodate the new location.
And since I was at it, also replaced the usage of hardcoded `/` with
`os.path.join()`, and changed from the `master` branch to `HEAD` which
always points to the current commit. The downside is that `HEAD` can
contain references, but the upside is that it now also works for tags
and all other commits.
This solution does not require deps like `gitpython` or assumes the
`git` command is avilable (though that would be resonable).