Slightly reorganize about dialogue
Moved the library versions to its own tab, so we have more space for
more and it can be easily copied. And moved the version logic into a
separate function, in the hope we can reuse that for e.g. #552.
Also moved the changeset to the line below the puddletag version one.
I intentionally did not update the translation resources, since they
are minor and there are already translation rework efforts going on.