Method QAbstractItemModel.reset() obsolete
It is replaced by beginResetModel() and endResetModel(). These are
wrapped in a python context manager, so model-resetting code can simply
surrounded with a `with` block and ensure that the end method is called
even in case of an error.
The comment regarding the signal is from 2009, I will *assume* this issue
is fixed by now, especially because endResetModel() will also emit the
modelReset signal.