Fix dirview refresh
In contrast to the old QDirModel, the new QFileSystemModel
automatically detects changes in the filesystem and updates itself.
Because of that a refresh is technically not needed, or even possible.
But you never know, so we keep it for now and call
`fetchMore(QModelIndex)` instead which should do something similar.

Fixes #708