Remove ghost window
The ghost window appeared because `setVisible(True)` was called on a
widget (`cancel` in this case) whose only parent was a layout (`cbox`).
This forced the button (and layout) to be displayed as/in a new window.
This window gets destroyed as soon as the layout gets a "proper"
parent which is not visible. This happens only a few lines after when
`vbox` is set as layout for the ProgressWin QDialog.
To solve this, do not set the button to explicitly visible but only
hide it if it shouldn't be displayed.
I am not going to question why the button is only hidden instead of
completely removed, I will just accept it as given right now.