Simplify package version
The version was and continues to be PEP-440 [0] compliant. The
intermediate list is removed to keep it as static as possible. The whole
changeset logic is guarded by a additional check, because the __file__
global may not always be available. While PEP-396 [1] is considered
deprecated, it is still a common convention and compatibility is added.

[0] https://peps.python.org/pep-0440/
[1] https://peps.python.org/pep-0396/