Make build PEP-517 compliant
The `setup.py` needed to be adjusted to allow building using a PEP-517
tool like pypa-build.
Simply importing the `puddlestuff` package does not work, because the
package-to-be-build is not in the search path, which makes sense because
the runtime dependencies that may be needed aren't available either.
Instead we directly import/exec the file we need, which only needs the
stdlib. Put all that into a function to keep it contained.
Also added a minimum required version for python. PyQt5 requires 3.6 but
that is already EOL, so the next is 3.7.

Also made sure that direct calls to setuptools like
`python 'setup.py' 'sdist'` still work.