Rewire menu item to issue template
Instead of adding a new menu item I repurposed the existing "Bug tracker"
one, because right above it is the GitHub link/item and it is common
enough that you can assume a project hosted on GitHub will use its issue
system as bug tracker.
It amazes me that you need a 79 line function only to reliably get a
useful linux distribution name across a range of python versions. Not
really sure what all that up and down in the platform module is about.

Closes #552