Fix cover art for APEv2
The bytes-operations on read and write weren't properly migrated from
python v2 to v3, and thus failed. Existing cover art was discarded on
read, and an exception thrown on write; both work now.

The restriction to only delete APEValues with kind 0 (Text) means it
was impossible to delete cover art which has kind 1 (Binary). Not really
sure what the reason for that restriction was, but it's from the time
before cover art was supported. External values are still excluded
since I am not sure how they work in puddletag (if at all).

Fixes #663