Make sorting locale aware
The case-sensitivity behaves a bit weird, because it only changes the
sorting of strings that only differ in case.
Case-sensitive will always put the lowercase variant first.
[AC,ab,aa,AA,ac] becomes [aa,AA,ab,ac,AC], aa and ac come before AA and
AC respectively, but in general uppercase and lowercase is still mixed.
Case-insensitive works more or less the same, but instead of putting the
lowercase-variant always first it preserves the original order. The same
[AC,ab,aa,AA,ac] becomes [aa,AA,ab,AC,ac], aa and AC come before AA and
ac respecively because they were originally in that order.
On the other side, the only place where you can switch between the two
is the "sort fields" function.

Fixes #728