Migrate all icons to names
All icons across the whole codebase are only referred to by their name,
not their filename. This has several advantages:
- The icons can (and will) be looked up in the users icon theme,
  providing a better integrated look and feel.
- All icons, including the puddletag-specific ones, can easily be over-
  ridden/replaced by users with their icon theme (Hello dark mode
  people).
- No longer multiple inconsistent combinations of icon names and
  filesnames
- Consistency between the icon name and the filename

It shouldn't be necessary to include fallback icons for the standard
icons, since they are almost guaranteed to be provided by the users
icon theme anyway. But we did until now, so why not continue. Plus we
always need the puddletag-specific icons anyway.

Also renamed a few of the puddletag-specific icons to more clearly
convey their usage, and fit with the naming of icons (e.g. dash has a
special meaning).