Upgrade puddletag-specific icons
Replaced the black text icons with self-made compounds from existing
oxygen icons.

autonumber is document-multiple with a blue one on top.
filetotag is audio-x-generic next to document-import.
textfiletotag is text-plain next to document-import.
renamedirs is document-export next to folder.
tagtofile is document-export next to audio-x-generic.

focustree is replaced with standard view-sidetree.