Fix shortcuts config version
The `shortcuts` config file has version 29, while the `menus` config as
well as the `__version__` in `loadshortcuts.py` it is compared with are
already at 30 (bumped in commit 7c75da3d). This means the shortcuts
config file is currently always overwritten because it always fails the
version check.

Also moved the version from somewhere in the middle to the top of the
file, so it is easier to find and edit.