Migrate config data
The config files here use the superfluous `./` in the path. The reason
for that is that puddletag does some path-magic using `os.path` which
relies on the fact that the path of the resource have a `/` right before
the actual filename. I don't want to touch that right now, so using this
workaround to keep it working for now.

Also remove the old ini-style and by now unused puddletag.conf.