Move translations
There is a slight name clash since there already is a `translations.py`,
but since this is just a folder and not a package it should be fine.

This also includes two fixes: afr was missing in the .pro file, so it
wasn't updated by the update_translation.py script. And the unused and
practically empty es_es translation is removed, in favour of the better
and already used es_ES one.