Fix preview of filepath changes
There are multiple fields that store different parts of the filepath,
like __filename and __dirname. With a real MockTag (like in non-preview
mode) these fields are passed to the matching properties of the MockTag
class, which in turn all (sometimes with some indirection) read/write
from/to a single attribute named __filepath, which contains the absolute
path of the file. This also means that writing one of the fields will
automatically update all the others.
The preview doesn't have these properties/attribute, it is just a dict.
And it wasn't treating these fields special, and didn't update the
related fields. This is the reason why for example the __filename field,
which is displayed in the table or tag panel, still showed the original
value after you used the "tag to filename" function which only set the
__path field. As soon as the preview is written, the MockTag implemen-
tation is used again, and the fields are properly updated again.

This implements the automatic update of the filepath fields in the
preview, so it will always show what the result will be.

The remove_from_preview function was added in 5862184b but never used,
so removed the dead code.