Fix delete file segfault
The worker function calls a non-thread-safe method of the model, which
can result in a segfault. Qt provides "queued connections" for these
cases to allow making such calls in a thread-safe way. Signals are one
way to employ them, but they require a bit more boilerplate; for example
a QObject instance to emit them. It is easier in this case to make the
function call using `invokeMethod`, which will automatically "do the
right thing" to make that call in a thread-safe way. While signals work
with implicit slots, in this case you need an explicit one which can
simply be done with an decorator.

For more info see https://wiki.qt.io/Threads_Events_QObjects